feat: add cursor rules and update import paths for categorizer, feed, and paper repositories
