+++ app/service/services/sync/pollyfills/paper.ts
@@ -1,311 +1,11 @@
 import { Entity } from "@/models/entity";
-import {
-  zPaper,
-  Paper as SqlitePaper,
-  Feed as SqliteFeed,
-  zFeed, zPaperFieldVersion,
-  zFeedFieldVersion,
-  Library as SqliteLibrary,
-  zLibrary,
-  zTag,
-  zTagFieldVersion,
-  zFolder,
-  zFolderFieldVersion
-} from "@/service/services/database/sqlite/models";
+import { zPaper, zPaperFieldVersion, Paper as SqlitePaper } from "@/service/services/database/sqlite/models";
 import { syncStateStore } from "@/service/services/sync/states";
-import z from "zod";
 import { db } from "@/service/services/database/sqlite/db";
 import { v4 as uuidv4 } from 'uuid';
-import { Feed } from "@/models/feed";
-import { CategorizerType, ICategorizerObject } from "@/models/categorizer";
-
-/**
- * Creates a structured value object for field versions that includes operation type,
- * previous value, and new value for complete audit trail.
- */
-function createFieldVersionValue(operation: 'create' | 'update', previousValue: any, newValue: any): string {
-  return JSON.stringify({
-    operation,
-    previousValue: previousValue ?? null,
-    newValue: newValue ?? null,
-  });
-}
-
-/**
- * Converts undefined values to null for SQLite compatibility.
- * SQLite cannot handle undefined values but can handle null.
- */
-function ensureUndefinedToNull(value: any): any {
-  return value === undefined ? null : value;
-}
-
-/**
- * Converts boolean values to integers (0 or 1) for SQLite compatibility.
- * SQLite cannot bind boolean values directly, so we convert them to integers.
- * @param value - The boolean value to convert
- * @returns 1 for true, 0 for false, null for null/undefined
- */
-function booleanToInt(value: boolean | null | undefined): number | null {
-  if (value === null || value === undefined) {
-    return null;
-  }
-  return value ? 1 : 0;
-}
-
-/**
- * Converts integer values back to booleans for application use.
- * @param value - The integer value to convert (0, 1, or null)
- * @returns true for 1, false for 0, null for null
- */
-function intToBoolean(value: number | null | undefined): boolean | null {
-  if (value === null || value === undefined) {
-    return null;
-  }
-  return value === 1;
-}
-
-/**
- * Ensure the library ID in sqlite database. If the library is not in database, insert it to database.
- * Otherwise, return the library ID.
- * @param library The library name. If not provided, the default library "main" will be used.
- * @returns The library ID. Guaranteed in database.
- */
-export async function ensureLibraryId(library?: string): Promise<string> {
-  const deviceId = syncStateStore.get("deviceId");
-
-  // Try get the existed sqlite library by legacy oid
-  const existedSqliteLibrary: SqliteLibrary | undefined = await db.selectFrom("library")
-    .where("name", "=", library ?? "main")
-    .selectAll()
-    .executeTakeFirst();
-
-  if (!existedSqliteLibrary) {
-    const newSqliteLibrary: z.infer<typeof zLibrary> = {
-      id: uuidv4(),
-      name: library ?? "main",
-      description: null,
-      ownedBy: syncStateStore.get("userId"),
-      createdAt: new Date().toISOString(),
-      updatedAt: new Date().toISOString(),
-      createdByDeviceId: deviceId,
-      updatedByDeviceId: deviceId,
-      deletedAt: null,
-      deletedByDeviceId: null,
-    }
-    await db.insertInto("library").values(newSqliteLibrary).execute();
-    return newSqliteLibrary.id;
-  }
-  return existedSqliteLibrary.id;
-}
-
-
-/**
- * Convert the feed to sqlite object. If the feed is not in database, insert it to database.
- * Otherwise, update the database feed according to the feed.
- * Finally, return the sqlite object.
- * @param feed The feed to be converted to sqlite object.
- * @returns The sqlite object. Guaranteed in database.
- */
-export async function toSqliteFeed(feed: Feed, library?: string): Promise<z.infer<typeof zFeed>> {
-  const deviceId = syncStateStore.get("deviceId");
-
-  // Try get the existed sqlite feed by legacy oid
-  const existedSqliteFeed: SqliteFeed | undefined = await db.selectFrom("feed")
-    .where("legacyOid", "=", feed._id.toString())
-    .selectAll()
-    .executeTakeFirst();
-
-  if (existedSqliteFeed) {
-    let updated = false;
-    const createdAtDate = new Date();
-    const createdAtDateString = createdAtDate.toISOString();
-    const createdAtTimestamp = createdAtDate.getTime();
-    // Update the sqlite feed if any field is different
-    const feedFieldVersions: z.infer<typeof zFeedFieldVersion>[] = [];
-    if (existedSqliteFeed.name !== feed.name) {
-      existedSqliteFeed.name = feed.name;
-      updated = true;
-      feedFieldVersions.push({
-        id: uuidv4(),
-        feedId: existedSqliteFeed.id,
-        field: "name",
-        value: createFieldVersionValue('update', existedSqliteFeed.name, feed.name),
-        timestamp: createdAtTimestamp,
-        deviceId: deviceId,
-        hash: null,
-        createdAt: createdAtDateString,
-        createdByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      });
-    }
-    if (existedSqliteFeed.count !== feed.count) {
-      existedSqliteFeed.count = feed.count;
-      updated = true;
-      feedFieldVersions.push({
-        id: uuidv4(),
-        feedId: existedSqliteFeed.id,
-        field: "count",
-        value: createFieldVersionValue('update', existedSqliteFeed.count, feed.count),
-        timestamp: createdAtTimestamp,
-        deviceId: deviceId,
-        hash: null,
-        createdAt: createdAtDateString,
-        createdByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      });
-    }
-    if (existedSqliteFeed.color !== feed.color) {
-      existedSqliteFeed.color = ensureUndefinedToNull(feed.color);
-      updated = true;
-      feedFieldVersions.push({
-        id: uuidv4(),
-        feedId: existedSqliteFeed.id,
-        field: "color",
-        value: createFieldVersionValue('update', existedSqliteFeed.color, feed.color),
-        timestamp: createdAtTimestamp,
-        deviceId: deviceId,
-        hash: null,
-        createdAt: createdAtDateString,
-        createdByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      });
-    }
-    if (existedSqliteFeed.url !== feed.url) {
-      existedSqliteFeed.url = feed.url;
-      updated = true;
-      feedFieldVersions.push({
-        id: uuidv4(),
-        feedId: existedSqliteFeed.id,
-        field: "url",
-        value: createFieldVersionValue('update', existedSqliteFeed.url, feed.url),
-        timestamp: createdAtTimestamp,
-        deviceId: deviceId,
-        hash: null,
-        createdAt: createdAtDateString,
-        createdByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      });
-    }
-    if (updated) {
-      existedSqliteFeed.updatedAt = new Date().toISOString();
-      existedSqliteFeed.updatedByDeviceId = deviceId;
-      await db.insertInto("feedFieldVersion").values(feedFieldVersions).execute();
-      await db.updateTable("feed").set(existedSqliteFeed).where("id", "=", existedSqliteFeed.id).execute();
-    }
-
-    return existedSqliteFeed;
-  }
-  // Insert the feed to database
-  const createdAtDate = new Date();
-  const createdAtDateString = createdAtDate.toISOString();
-  const createdAtTimestamp = createdAtDate.getTime();
-  const newSqliteFeed: z.infer<typeof zFeed> = {
-    id: uuidv4(),
-    legacyOid: ensureUndefinedToNull(feed._id?.toString()),
-    name: feed.name,
-    description: null,
-    ownedBy: ensureUndefinedToNull(syncStateStore.get("userId")),
-    libraryId: await ensureLibraryId(library),
-    type: "rss",
-    url: feed.url,
-    count: ensureUndefinedToNull(feed.count),
-    color: ensureUndefinedToNull(feed.color),
-    createdAt: createdAtDateString,
-    createdByDeviceId: deviceId,
-    updatedAt: null,
-    updatedByDeviceId: null,
-    deletedAt: null,
-    deletedByDeviceId: null,
-  }
-  const feedFieldVersions: z.infer<typeof zFeedFieldVersion>[] = [
-    {
-      id: uuidv4(),
-      feedId: newSqliteFeed.id,
-      field: "name",
-      value: ensureUndefinedToNull(createFieldVersionValue('create', null, feed.name)),
-      timestamp: createdAtTimestamp,
-      deviceId: deviceId,
-      createdAt: createdAtDateString,
-      createdByDeviceId: deviceId,
-      hash: null,
-      deletedAt: null,
-      deletedByDeviceId: null,
-    },
-    {
-      id: uuidv4(),
-      feedId: newSqliteFeed.id,
-      field: "type",
-      value: ensureUndefinedToNull(createFieldVersionValue('create', null, "rss")),
-      timestamp: createdAtTimestamp,
-      deviceId: deviceId,
-      createdAt: createdAtDateString,
-      createdByDeviceId: deviceId,
-      hash: null,
-      deletedAt: null,
-      deletedByDeviceId: null,
-    },
-    {
-      id: uuidv4(),
-      feedId: newSqliteFeed.id,
-      field: "url",
-      value: ensureUndefinedToNull(createFieldVersionValue('create', null, feed.url)),
-      timestamp: createdAtTimestamp,
-      deviceId: deviceId,
-      createdAt: createdAtDateString,
-      createdByDeviceId: deviceId,
-      hash: null,
-      deletedAt: null,
-      deletedByDeviceId: null,
-    },
-    {
-      id: uuidv4(),
-      feedId: newSqliteFeed.id,
-      field: "count",
-      value: ensureUndefinedToNull(createFieldVersionValue('create', null, feed.count)),
-      timestamp: createdAtTimestamp,
-      deviceId: deviceId,
-      createdAt: createdAtDateString,
-      createdByDeviceId: deviceId,
-      hash: null,
-      deletedAt: null,
-      deletedByDeviceId: null,
-    },
-    {
-      id: uuidv4(),
-      feedId: newSqliteFeed.id,
-      field: "color",
-      value: ensureUndefinedToNull(createFieldVersionValue('create', null, feed.color)),
-      timestamp: createdAtTimestamp,
-      deviceId: deviceId,
-      createdAt: createdAtDateString,
-      createdByDeviceId: deviceId,
-      hash: null,
-      deletedAt: null,
-      deletedByDeviceId: null,
-    },
-  ]
-  await db.insertInto("feed").values(newSqliteFeed).execute();
-  await db.insertInto("feedFieldVersion").values(feedFieldVersions).execute();
-
-  return newSqliteFeed;
-}
-
-export async function deleteSqliteFeed(legacyOid: string): Promise<void> {
-  const deviceId = syncStateStore.get("deviceId");
-  // Query and get the sqlite feed by legacy oid
-  const existedSqliteFeed = await db.selectFrom("feed").where("legacyOid", "=", legacyOid).selectAll().executeTakeFirst();
-  if (existedSqliteFeed) {
-    await db.updateTable("feed").set({
-      deletedAt: new Date().toISOString(),
-      deletedByDeviceId: deviceId,
-    }).where("id", "=", existedSqliteFeed.id).execute();
-    await db.deleteFrom("feedFieldVersion").where("feedId", "=", existedSqliteFeed.id).execute();
-  }
-}
+import z from "zod";
+import { createFieldVersionValue, ensureUndefinedToNull, ensureLibraryId, booleanToInt } from "./utils";
+import { toSqliteFeed } from "./feed";
 
 
 /**
@@ -316,7 +16,7 @@ export async function deleteSqliteFeed(legacyOid: string): Promise<void> {
  * @param entity The entity to be converted to sqlite object.
  * @returns The sqlite object. Guaranteed in database.
  */
-export async function toSqliteEntity(entity: Entity, logService?: any): Promise<z.infer<typeof zPaper>> {
+export async function toSqlitePaper(entity: Entity, logService?: any): Promise<z.infer<typeof zPaper>> {
 
   const deviceId = syncStateStore.get("deviceId");
 
@@ -1324,7 +1024,7 @@ export async function toSqliteEntity(entity: Entity, logService?: any): Promise<
   return sqliteEntity;
 }
 
-export async function deleteSqliteEntity(legacyOid: string): Promise<void> {
+export async function deleteSqlitePaper(legacyOid: string): Promise<void> {
   const deviceId = syncStateStore.get("deviceId");
   // Query and update deletedAt and deletedByDeviceId of the sqlite entity by legacy oid
   const existedSqliteEntity = await db.selectFrom("paper").where("legacyOid", "=", legacyOid).selectAll().executeTakeFirst();
@@ -1361,249 +1061,5 @@ export async function deleteSqliteEntity(legacyOid: string): Promise<void> {
       deletedAt: deletedAtDateString,
       deletedByDeviceId: deviceId,
     }).where("id", "in", existedSqliteFolders.map((folder) => folder.id)).execute();
-  }
-}
-
-/**
- * Sync the categorizer to the sqlite database
- * @param categorizer - The categorizer to sync
- * @param type - The type of the categorizer
- * @returns The sqlite categorizer, guaranteed in database.
- */
-export async function toSqliteCategorizer(
-  categorizer: ICategorizerObject,
-  type: CategorizerType
-): Promise<z.infer<typeof zTag> | z.infer<typeof zFolder>> {
-  const deviceId = syncStateStore.get("deviceId");
-
-  if (type === CategorizerType.PaperTag) {
-    const createdAtDate = new Date();
-    const createdAtDateString = createdAtDate.toISOString();
-    const createdAtTimestamp = createdAtDate.getTime();
-    // Try get the existed sqlite tag by legacy oid
-    const existedSqliteTag = await db.selectFrom("tag").where("name", "=", categorizer.name).selectAll().executeTakeFirst();
-    const tagId = existedSqliteTag?.id || uuidv4();
-    // If the tag is not existed, insert it to database
-    if (!existedSqliteTag) {
-      const sqliteTag: z.infer<typeof zTag> = {
-        id: tagId,
-        name: categorizer.name,
-        description: null,
-        colour: ensureUndefinedToNull(categorizer.color),
-        createdAt: createdAtDateString,
-        createdByDeviceId: deviceId,
-        updatedAt: createdAtDateString,
-        updatedByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      };
-      await db.insertInto("tag").values(sqliteTag).execute();
-      return sqliteTag;
-    }
-
-    // If the tag is already existed, update the sqlite tag if any difference
-    let updated = false;
-    const updatedTimestamp = createdAtTimestamp;
-    const updatedDateTime = createdAtDateString;
-    const sqliteTagVersions: z.infer<typeof zTagFieldVersion>[] = [];
-    if (existedSqliteTag.name !== categorizer.name) {
-      existedSqliteTag.name = categorizer.name;
-      updated = true;
-      existedSqliteTag.updatedAt = updatedDateTime;
-      existedSqliteTag.updatedByDeviceId = deviceId;
-      sqliteTagVersions.push({
-        id: uuidv4(),
-        tagId: tagId,
-        field: "name",
-        value: createFieldVersionValue('update', existedSqliteTag.name, categorizer.name),
-        timestamp: updatedTimestamp,
-        deviceId: deviceId,
-        createdAt: updatedDateTime,
-        createdByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      });
-    }
-    if (existedSqliteTag.colour !== categorizer.color) {
-      existedSqliteTag.colour = categorizer.color;
-      updated = true;
-      existedSqliteTag.updatedAt = updatedDateTime;
-      existedSqliteTag.updatedByDeviceId = deviceId;
-      sqliteTagVersions.push({
-        id: uuidv4(),
-        tagId: tagId,
-        field: "colour",
-        value: createFieldVersionValue('update', existedSqliteTag.colour, categorizer.color),
-        timestamp: updatedTimestamp,
-        deviceId: deviceId,
-        createdAt: updatedDateTime,
-        createdByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      });
-    }
-    if (updated) {
-      await db.updateTable("tag").set(existedSqliteTag).where("id", "=", tagId).execute();
-      await db.insertInto("tagFieldVersion").values(sqliteTagVersions).execute();
-    }
-
-    return existedSqliteTag;
-  }
-  if (type === CategorizerType.PaperFolder) {
-    // Try get the existed sqlite folder by legacy oid
-    const existedSqliteFolder = await db.selectFrom("folder").where("name", "=", categorizer.name).selectAll().executeTakeFirst();
-    const folderId = existedSqliteFolder?.id || uuidv4();
-    // If the folder is not existed, insert it to database
-    if (!existedSqliteFolder) {
-      const createdAtDate = new Date();
-      const createdAtDateString = createdAtDate.toISOString();
-      const createdAtTimestamp = createdAtDate.getTime();
-      const sqliteFolder: z.infer<typeof zFolder> = {
-        id: folderId,
-        legacyOid: categorizer._id.toString(),
-        name: categorizer.name,
-        colour: ensureUndefinedToNull(categorizer.color),
-        description: null,
-        parentId: null,
-        createdAt: createdAtDateString,
-        createdByDeviceId: deviceId,
-        libraryId: await ensureLibraryId(),
-        updatedAt: createdAtDateString,
-        updatedByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      };
-      await db.insertInto("folder").values(sqliteFolder).execute();
-      // TODO: handle the children
-      return sqliteFolder;
-    }
-    // If the folder is already existed, update the sqlite folder if any difference
-    let updated = false;
-    const updatedAt = new Date();
-    const updatedAtString = updatedAt.toISOString();
-    const updatedTimestamp = updatedAt.getTime();
-    const sqliteFolderVersions: z.infer<typeof zFolderFieldVersion>[] = [];
-    if (existedSqliteFolder.name !== categorizer.name) {
-      existedSqliteFolder.name = categorizer.name;
-      updated = true;
-      existedSqliteFolder.updatedAt = updatedAtString;
-      existedSqliteFolder.updatedByDeviceId = deviceId;
-      sqliteFolderVersions.push({
-        id: uuidv4(),
-        folderId: folderId,
-        field: "name",
-        value: createFieldVersionValue('update', existedSqliteFolder.name, categorizer.name),
-        timestamp: updatedTimestamp,
-        deviceId: deviceId,
-        createdAt: updatedAtString,
-        createdByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      });
-    }
-    if (existedSqliteFolder.colour !== categorizer.color) {
-      existedSqliteFolder.colour = categorizer.color;
-      updated = true;
-      existedSqliteFolder.updatedAt = updatedAtString;
-      existedSqliteFolder.updatedByDeviceId = deviceId;
-      sqliteFolderVersions.push({
-        id: uuidv4(),
-        folderId: folderId,
-        field: "colour",
-        value: createFieldVersionValue('update', existedSqliteFolder.colour, categorizer.color),
-        timestamp: updatedTimestamp,
-        deviceId: deviceId,
-        createdAt: updatedAtString,
-        createdByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      });
-    }
-    if (updated) {
-      await db.updateTable("folder").set(existedSqliteFolder).where("id", "=", folderId).execute();
-      await db.insertInto("folderFieldVersion").values(sqliteFolderVersions).execute();
-    }
-    // As the update would not affact the id and parentId, we don't need to update the children
-    // TODO: handle the children update
-    return existedSqliteFolder;
-  }
-
-  // If the categorizer is not a tag or folder, throw an error
-  throw new Error(`Unknown categorizer type: ${type}`);
-}
-
-
-
-export async function deleteSqliteTag(name: string): Promise<void> {
-  const deviceId = syncStateStore.get("deviceId");
-  const existedSqliteTag = await db.selectFrom("tag").where("name", "=", name).selectAll().executeTakeFirst();
-  if (existedSqliteTag) {
-    await db.updateTable("tag").set({
-      deletedAt: new Date().toISOString(),
-      deletedByDeviceId: deviceId,
-    }).where("id", "=", existedSqliteTag.id).execute();
-    // Also delete the tagFieldVersion
-    await db.deleteFrom("tagFieldVersion").where("tagId", "=", existedSqliteTag.id).execute();
-    // Also add a delete operation to the paperTag
-    const existedPaperTag = await db.selectFrom("paperTag").where("tagId", "=", existedSqliteTag.id).selectAll().executeTakeFirst();
-    if (existedPaperTag) {
-      const createdAtDate = new Date();
-      const createdAtDateString = createdAtDate.toISOString();
-      const createdAtTimestamp = createdAtDate.getTime();
-      await db.insertInto("paperTag").values({
-        id: uuidv4(),
-        createdAt: new Date().toISOString(),
-        op: "remove",
-        timestamp: createdAtTimestamp,
-        deviceId: deviceId,
-        paperId: existedPaperTag.paperId,
-        tagId: existedPaperTag.tagId,
-        createdByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      }).execute();
-    }
-  }
-}
-
-export async function deleteSqliteFolder(legacyOid: string): Promise<void> {
-  const deviceId = syncStateStore.get("deviceId");
-  const existedSqliteFolder = await db.selectFrom("folder").where("legacyOid", "=", legacyOid).selectAll().executeTakeFirst();
-  if (existedSqliteFolder) {
-    await db.updateTable("folder").set({
-      deletedAt: new Date().toISOString(),
-      deletedByDeviceId: deviceId,
-    }).where("id", "=", existedSqliteFolder.id).execute();
-    // Also delete the folderFieldVersion
-    await db.deleteFrom("folderFieldVersion").where("folderId", "=", existedSqliteFolder.id).execute();
-    // Also add a delete operation to the paperFolder
-    const existedPaperFolder = await db.selectFrom("paperFolder").where("folderId", "=", existedSqliteFolder.id).selectAll().executeTakeFirst();
-    if (existedPaperFolder) {
-      const createdAtDate = new Date();
-      const createdAtDateString = createdAtDate.toISOString();
-      const createdAtTimestamp = createdAtDate.getTime();
-      await db.insertInto("paperFolder").values({
-        id: uuidv4(),
-        createdAt: createdAtDateString,
-        op: "remove",
-        timestamp: createdAtTimestamp,
-        deviceId: deviceId,
-        paperId: existedPaperFolder.paperId,
-        folderId: existedPaperFolder.folderId,
-        createdByDeviceId: deviceId,
-        deletedAt: null,
-        deletedByDeviceId: null,
-      }).execute();
-    }
-  }
-}
-
-
-export async function deleteSqliteCategorizer(legacyOid: string, type: CategorizerType): Promise<void> {
-  if (type === CategorizerType.PaperTag) {
-    await deleteSqliteTag(legacyOid);
-  }
-  if (type === CategorizerType.PaperFolder) {
-    await deleteSqliteFolder(legacyOid);
   }
 }
